gitconfig: bypass git-icdiff script on `dd` alias
This updates the `dd` alias to *not* call the git-icdiff script.  Instead
it directly invokes the cmd that the script would have.  This means
you don't have to mess with all the pager logic the script does and
instead lets git handle pagination.  This makes for consistent paging
between the `d` and `dd` aliases.

This also explicitly sets the `core.pager` to `less`.