updated ps1: include current time; dollar sign on newline
The main thing is having the meta on one line and the dollar sign
and command on its own line.  This makes seeing what the command
is easier and makes commands wrap lines less frequently.  Plus it
juxtaposes the pwd next to the command input - useful when cd'ing.

This other big change is adding the current time.  This is 1) so I
will be more aware of the time and 2) be a useful reference for when
I ran that command.

The other stuff is just minor cleanups:

* put the git info right next to the pwd - corresponds directly to
  the pwd
* remove the brackets from the ruby version - less clutter - the
  color is enough to mentally section it off

This was inspired by the ps1 used in this tutorial video:
https://www.youtube.com/watch?v=tS4yHxkG0K4